Ensure inverse document relations are always destroyed.
Although RelatedDocuments#published_related_documents is only intended
as a read-only method, Rails will allow it to be used to remove document
relations, so it seems sensible to add the same before_remove callback.
I tried to extract the duplication into a method, but for some reason
the method is only passed one of the documents, so I ended up extracting
the duplication into a constant. Not sure whether the method behaviour
described above is a bug in Rails or something I've missed.