Move some scenarios around for consistency
A distinction exists for some functionality between the "drafting" features, which exercise the admin interface, and the "viewing" and "publishing" features, which generate data using factories. However, some feature files contained both kinds of scenario, making it less clear whether or not it was appropriate to exercise the admin interface. Hopefully this is clearer.

I also spotted two duplicate scenarios between the files, which have now been removed.