Skip slimmer with rack middleware in test environment.
I wanted to test a routing error scenario, but this means that the
response is generated without having been through any Rails controller
and consequently without having had the skip slimmer response header
set. This meant that instead of the expected routing error, I was seeing
a SocketError raised by slimmer.

By adding this rack middleware in the test environment, we ensure that
the skip slimmer response header is set and the expected routing error
is raised.

This change is only really relevant for Rails integration tests and
Cucumber tests which test the whole stack.