Do not allow access to admin pages from single domain URLs.
We don't want the general public to have access to the admin part of the
web application. This should work in both preview & production
environments.

Note: we will need to change the configuration for this if/when the
single domain changes from alphagov.co.uk. It might be safer to reverse
the logic and check for specific hosts for admin pages, but this would
make it awkward to run the app locally in a development environment and
in tests, so I think this is the best option for now.