Create thumbnails from PDFs.
We can use ImageMagick's `convert` command to create a thumbnail of the first page of uploaded PDFs.

The actual conversion process is quite simple (although it would be interesting to find PDF instances which `convert` couldn't handle), but it too a lot of time to figure out what was necessary to save a carrierwave version with a different file extension. I am not convinced that carrierwave is the right choice here, particularly as I start to think about generating the thumbnails offline.