Ensure the thumbnails have the right extension.
Well.

Where to start.

CarrierWave really isn't very clear about how to manipulate the filename for versions of an attachment. I had thought that `filename` was working, and it certainly did appear to create the right file when processing the thumbnail, but it didn't result in the right URL being generated.

After a literal day of investigating the guts of the library, I believe I have determined that overriding the `full_filename` method will ensure that the uploaded file is created with the right extension, and that the same filename is used when generating the URL for the file.

Without this change, the cucumber feature will fail (because the ".png" extension is missing from the URL). However, the unit test will still pass. This appears to be some quirk with CarrierWave behaviour slightly different in different situations. Confusing? Yes. Annoying? VERY.