Better thumbnail styling.
If the attachment is a PDF, then display the thumbnail to one side, and the details by its side.

I've floated the thumbnail, which required setting overflow: auto on both the attachment and document container divs. I don't think this has any adverse affect, but if anyone knows of one I'd be happy to do something else.