Editors can feature published consultations.
The tests for displaying the 'feature' and 'unfeature' buttons are
necessarily in the documents_controller_test as the documents_controller
is responsible for rendering the document lists. The tests for featuring
and unfeaturing a consultation are just in the
consultations_controller_test as I've only made the actions available on
that controller at the moment.

This featuring/unfeaturing mechanism is different to that for news
articles where you set the 'featured' flag when editing the article.
I'll change that to use this mechanism in a future commit. This will
bring the 'featuring' mechanism inline for policy areas, consultations
and news articles.