News articles can be featured from the document list.
This makes the featuring of news articles consistent with the featuring
of consultations and policy areas.

I've extracted/generalised the consultation controller tests and re-used
them in the news article controller test.

I've removed, rather than replaced, the 'Creating a featured news
article' cucumber scenario as I'm not sure it was testing much about the
outcome of featuring a news article anyway.