Authenticate using the single-sign-on service.
Firstly, apologies for the huge size of this commit. I'll try and outline the major changes.

The GDS-SSO mechanism authenticates via OAuth2, and then attempts to locate the corresponding user in our database by UID. If that user is not found, it will create a new user matching those details. We're going to need to do some work to ensure that the new sign-on-o-tron users are mapped to the correct existing users in our database.

Unfortunately, there doesn't seem to be a nice way to stub gds-sso's usage of Warden at the moment. The provided test mechanisms only operate in full-stack (integration/cucumber) tests, so functional tests use a warden stubbing strategy, and integration tests use an enhancement to gds-sso that allows the stubbing of specific users.

I've had to remove quite a lot of our custom tests, along with our session controller and routes, and replace them with fairly flimsy tests that assure unauthenticated users are at least sent in the right direction for authentication (i.e. the redirect to /auth/gds). It would be great to test this in a more comprehensive way.