Documents can only be related to policies.
Previously, documents could be arbitrarily related, but this has not been true for some time.

There is a bug where editing a draft policy will remove associations from other documents to this policy. This ultimately stems from `Policy` including the `Document::RelatedDocuments` module, which then prompts the `DocumentController` to wipe the related documents upon saving the `Policy`.

Ultimately I need to remove the inclusion of this module from `Policy`, but before we get there I think we can clarify the nature of the related documents association.