Extract duplicate tests into test helper.
I intentionally didn't extract duplication from
Admin::SpeechesControllerTest in
57f3972e4aaffef06390f1d030206d3408fdfb05, because of the complication
with a speech factory not providing all the necessary attributes to
successfully create a speech using the create controller action. This
commit resolves that problem by introducing a
`controller_attributes_for` method in the test helper which is
over-ridden in the speeches controller test.