Document relationships shouldn't be cleared when policies are saved.
They were being destroyed because the the controller assumed that any update for a class which claimed that it could be associated with related documents would also include the set of related document IDs. This isn't true for `Policy` instances.

I also had to refactor some of the document tests, and took the opportunity to re-orient the relationship around the non-policy documents; they are the instances where the relationship is really maintained.