Remove `should_display_attachments_for` admin helper.
The behaviour of displaying attachments in the admin interface is
closely tied to the behaviour of uploading attachments so I've moved all
the tests into the `should_allow_attachments_for` admin helper.

This removes the overriding of the `should_display_attachments_for`
helper method (previously defined in both
`DocumentControllerTestHelpers` and
`AdminDocumentControllerTestHelpers`).