Make summary appear above body.
Summary is now available in the standard fields, but only visible for those types of document that already had it. However, it should now be very easy to add summaries to any new documents.

This got rid of a bunch of duplication, and standardised the size of the summary input.