Remove our use of STI for speech types.
I've wanted to do this for a while but came up with a fairly valid
excuse while investing the performance of the site. Consider the
following:

    class SpeechController
      def index
        @speeches = Speech.all.includes(:document_identity)
      end
    end

    in speeches/index.html.erb
    <% @speeches.each do |speech| %>
      <%= link_to speech.title,
          public_document_path(speech.becomes(Speech)) %>
    <% end %>

The presence of `#becomes` in the template results in ActiveRecord
issuing a query to find the `DocumentIdentity` even though they were
pre-loaded in the controller.

I'm hoping this change will allow us to remove some other special case
code around Speeches too.