Avoid n+1 queries on /consultations.
I narrowed down the relationships that needed to be included by
monitoring the development log file while making the relevant changes in
the controller.

I've added mock tests that will fail fast if the things that are
included change. That's not to say that the inclusions shouldn't change,
just that it should be a conscious decision if they do.

The tests won't prevent us introducing more n+1 queries in future.