Make Whitehall responsible for displaying search results.
This replaces our public version of rummager with a search controller in
our app. The SearchClient requests the search results from rummager and
we render them in our own template.

We've previously done quite a lot of work generalising Rummager so that
we can host a version for Whitehall. This commit renders quite a lot of
that unnecessary which makes me feel a little uncomfortable. However,
I'm not sure it's worth continuing down the generalisation approach when
we can do what we need easily in our own app.