Allow supporting documents to have attachments.
I had a few options about how to implement this, including adding STI to Attachment, or converting SupportingDocument to a Document subclass, but in the end this seemed like the simplest way. Some partials have been shared with Document even though SupportingDocument is not strictly a Document subclass; where any parameterisation of partials was required, I have simply duplicated the body for later refactoring.