Populate missing DocumentIdentity#document_types
Commit f8ed51c9d5b80552080cf7917ffafe7866a57023 added the document_type
field to document_identities so that we can have different document
types with the same slug. Unfortunately, we didn't update existing
document identities at that time which resulted in a number of documents
returning 404s (because the query tries to find them based on slug and
document_type and the document_type is NULL).

This migration should populate the missing document types for those
document identities that currently have a NULL document_type.

*NOTE* As of this commit there are two records in production with a slug
of 'giving-communities-rights'. They are both policies yet one has an
empty document_type. Running this migration will fail because of the
unique constraint on slug and document_type. To avoid this problem I'm
going to manually update the later record to give it the slug it
should've got using the following SQL:

    UPDATE document_identities
    SET slug = "giving-communities-rights--2"
    WHERE id = 454;