Don't mess up UTF-8 characters.
I had to change the test covering the preservation of unicode characters when rewriting links; after some discussion with James Mead, I decided to simplify the test to specify only the expected output, rather than testing the implementation. This has the downside that it may only fail on CI (which uses the bugging libxml versin 2.7.3) and not locally, unfortunately.

I had hoped I could implement a change which would actually leave the HTML entities as they are (i.e. &lsquo; and &rsquo;), which is possible leaving out line 53. However, this breaks the actual UTF-8 handling during the re-write of links.