Don't highlight navigation on search results page
The previous behaviour was to highlight Home when on any "miscellaneous"
page, but search is currently the only page which doesn't fit naturally
under one of the main navigation items.

Now we don't highlight the navigation at all for "miscellaneous" pages;
if any are added in future and their navigation behaviour should be
different (e.g. highlighting Home), that'll need to be implemented
explicitly.