We don't need to sidestep validations when featuring.
It should be possible to feature a document that's already been published, but given there's an existing set of updatable attributes on a published document, lets augment that rather than sidestepping validation with `update_attribute`.

I've also modified the tests to actually use a published document, as the test names imply.

Extending this approach should make it easier to also update the featured image on a published document.