Collapse Featuring attributes back into Document.
Having these attributes in a separate model was complicating things and
although adding more columns to the documents table isn't ideal, it is
at least consistent with how we are dealing with this elsewhere in the
code. A more useful change would be to move these attributes onto
DocumentIdentity so we don't need to add more exceptions in the
Document#significant_changed_attributes method, but that can be a later
commit.