Try to import policy stubs against listed people
We don't want every cabinet office policy to be associated with David Cameron, for example.

In some cases this will still result in policies that aren't associated with the right ministerial roles, since when the person has multiple roles there's no way to tell which of those is correct (e.g. Theresa May)