Fix a problem with multiple attachment javascript.
When we copy the elements under the "file_upload" classed element, we
should make the values of the input elements and the contents of the
"already_uploaded" element blank to avoid weirdness after a validation
failure i.e. when a file is shown as "already uploaded".

Here are the steps to reproduce the bug :-

1. Create a new document that allows attachments (e.g. publication)
2. Choose a file as an attachment, but leave one or more mandatory fields blank (e.g. document title)
3. Click save
4. You will see errors marked against the blank field(s) and the attachment will be marked as "already uploaded"
5. Choose a different file for the attachment
6. If javascript is enabled, you will see a new attachment field appear which will also say it is "already uploaded". This is wrong.
7. If you now click save, both the old and the new file will be attached, but the second attachment should have been blank and only the newly selected file should have been uploaded.