Add mandatory title for attachments.
- The official title of a file attached to a publication page is
important information which we should not lose. We had been achieving
this by setting the publication page title to match the official title
of the most relevant attached file. However this has led to some very
long page titles which are not very suitable for e.g. list views of
publications.
- By introducing a title on each attachment we can capture the official
title of each attached file and free up the publication page title for
something a little snappier and more appropriate for list views.
- An additional benefit is that we may be able to use the attached file
title to improve accessibility for the "Download attachment" link.
- There were a few complications in introducing this extra field e.g.
the multiple_file_upload javascript now has to cope with the
"field_with_errors" wrapping divs which are added on validation failure.
- Titles for existing attachments have been generated fairly naively
based on the attachment filename.