Improve the usability of the organisation contact form.
We hide the empty contact fields and use them as a template for dynamically adding new contacts; empty fields are hidden by default but can be shown by clicking a menu that's fixed to the top right of the contact element. This should allow contacts of various kinds (just telephone numbers, just email address, just actual address, or any combination) to be added arbitrarily.

It's worth noting that you can't delete a contact at the moment; the best you can do is empty it of data but I doubt that's really sufficient. There's also no control over the ordering of the contacts, but this can theoretically be handled by copying the field values around manually.