Add change note to documents, but don't use it yet.
This is a bit of a cheat. I've set
Document::Publishing#change_note_required? to always return false for
now so as to reduce the size of my commit. I've actually implemented all
the necessary changes to make this work for real, but hopefully breaking
down the commits like this will make it easier to see what's going on.