Require change note for publishing all but first edition of document.
The change note is stored against the document edition, but not yet
displayed anywhere. I'm not entirely happy with this solution, but it's
the best I can come up with for now. Some things that are sub-optimal
are :-

- When a change note is required but not supplied, a flash alert message
appears, but the change note field is not highlighted as being the problem.
I'm going to attempt to address this in a subsequent commit.
- Usability-wise, it @jasoncale has suggested that when a change note is
required, it only appears after you've clicked publish, but it should be
possible to enhance the behaviour with javascript to achieve this.
- It would be nice to hook these publication rules more closely into the
document state machine. I may look at that in a subsequent commit.

Also note that some cucumber scenarios were a bit too robust in that they
relied on the publish step working and were asserting that something
stayed the same. These assertions would continue to pass even if the
publishing failed. I've tried to add checks that the publishing has worked
into these scenarios.