Highlight change note field when publishing subsequent editions.
As suggested by @jasoncale, use javascript to make change note field
more prominent. We hide the real form and insert a publishing button
link which when clicked shows the form and highlights the change note
field. Currently using the standard `field_with_errors` class for
highlighting - not sure how sensible that is.