Lock to Rake 0.9.2, rather than 0.9.2.2
We had originally allowed the use of 0.9.2.2 so that we could use Ruby
1.9.3 to develop against, but a change in rummager
(alphagov/rummager@c7171b056e92d5743beb971fdb96e4f99bd84047) locking
it to 0.9.2 explicitly, in combination with the fact that both apps
are deployed on the same machine, means that we have to revert this to
avoid Passenger freaking out.

Rake 0.9.2 is the version which puppet ensures, and so rather than try to retain this flexibility, it is easier to move Whitehall into line with the strict constraint.

Note that this means we can no longer develop using Ruby 1.9.3 locally.