Index supporting pages with their own format
These were previously masquerading as their associated document
in order to get the policy icon, but now there's an explicit
supporting page icon (albeit the same as the policy icon at the
moment) we can give them a distinct format.

If supporting pages ever get associated with other types of
document it might be desirable to revert this change. For now
it's better to keep things simple and unsurprising.