Stop news images appearing too small.
We're not exactly sure why this wasn't working, since `#content img`
has `width: 100%` as a rule, which should trigger it to be 100% of its
container.

The new rule is perhaps a bit more explicit; the `max-width` of the
image is set to 100% of its container, and the specific `width` of the
image is set to `inherit`, which should also set it to be the width of
the container.

Huh. Weird.