Avoid mass assignment warnings in the log.
We were attempting to copy the ID and type of a document into the new
draft document. Fortunately Rails was stopping this happening, but it's
better not to try and so to avoid the warnings in the log. The type is
effectively copied by the `self.class.new` call within
Document#create_draft.