Only allow revision of published editions.
- This was actually the intention of the original code, but because an
`else` was used instead of `elsif`, the result of the call to
`document.published?` was being ignored.
- There were no tests around when the `Create new edition` button
should not appear.
- I had to change a cucumber step that was just finding an existing
document by title and was not finding the latest edition. I've had to do
this in some previous commits and I feel it is (yet another) pointer
that we should be exposing the document identity model in the admin
interface instead of the document; or at least redirecting admin
document show pages to the latest edition.