Introduce Ranks for Ministerial Roles.
This is fundamentally to allow us to order the list of ministers that
appear on an organisation's ministers page. *NOTE* This commit itself
won't result in the correct ordering because we don't have any rank data
in the system yet. I haven't yet decided how best to populate the set of
ranks.

The `MinisterialRole.ordered_by_rank` method is slightly more
complicated than might be expected: We have to use a left outer join
(instead of an inner join) and coalesce the value of ranks.position.
Both of these are required to deal with the fact that, as of now, we
don't have any ranks in our system.

I've explicitly restricted the rank relationship to MinisterialRoles (as
opposed to BoardMemberRoles) to illustrate that they currently only
apply to MinisterialRoles.