Revert "Prevent access to admin functionality for all but latest edition."
This reverts commit 5f83fbde23946d9a5af3bcdaf0207098db0a0f8b.

It turns out that I had misunderstood how the admin interface worked.
You should in fact be able to *view* multiple editions of a single
document e.g. when one edition is published and another is a new draft.

9a680301accf1eca78208595b77e91abb7c29001 and
4913b49cc5d45d936c6e4d81e0e55c28482180e7 are probably sufficient to
prevent the bug we saw when a new edition was created based on an *old*
edition, thereby losing its supporting pages.