Add validation to Organisation Role.
We've ended up with some organisation_roles that don't have
corresponding roles. I haven't been able to replicate this through the
interface so my guess is that it used to be possible to remove a role
that was associated with an organisation. This change should at least
make it obvious if we try to remove a role that's associated with an
organisation in future.