Remove this test which checks for grid classes on non-div elements.
Hopefully now we aren't going to abuse the classes, for poor css choices, and it will ultimately mean we can have cleaner markup.

It's useful for example to have list items set in the grid, which cannot be surrounded by div elements and removes the need to add extra grid based declarations to their styles.