Use first image for news article featured on org page.
Pete pointed out that with the previous implementation a news article
needed to be feautured on the home page in order to add an image to be
displayed for an article featured only on an organisation page. This is
hopefully a better solution for now, although it does mean there's
another place where we need to worry about the aspect ratio of images.