Admin should be blocked for all requests arriving through the single domain
As well as adding a test, this commit also changes the previous blocked request test to use `get` in place of `get_via_redirect`.  It's important that if these requests get through to the server at all, that they should be blocked immediately.  Previously the authentication code would redirect to a 404, causing the test to pass but not for the reason we wanted.

Also, a corresponding test has been added to the smokey repository, which will be pushed as soon as this code has been deployed.