HMRC don't have a branding colour, so using Treasury branding until new colour chosen.
Nobody working today has been through the years of intensive training required to become a certified colour picker, so we are not qualified to choose a new colour ourselves.  A request for a new colour has been made, so soon, after an agonising process sifting through the millions of available colours to find the one that most represents HMRC, we'll be able to update this css file with the anointed hue.