Add an 'inside government' label to establish context.
In order to do this quickly in a way that doesn't break with the responsive header, we've changed the 'Latest' header link to 'Inside Govnernment'. We've also given it a bolder style than the other header navigation, to establish it more clearly.

We decided not to use orange as the colour because that signifies "beta", whereas this context will need to be present whether or not the site is in beta.