Clean up design on ministerial role page
There may be a better way to fix some of the issues I had doing this, but on the whole I'm pretty happy with how clean this has turned out.  The things I think might be a tiny bit wonky are:

a) To scope css rules, I added a .ministerial_role_page class to the whole content.  Using .ministerial_role seemed wrong as it is used elsewhere.
b) The header for news and speeches seemed had an unwanted left margin, so I zapped it.

Also, I just guessed the typography, putting something that looked alright rather than obsessing too much about it.