Change 'UK Embassy' label to 'UK contacts'.
'Some countries have a High Commission rather than an Embassy (eg Uganda). Some
UK in the world sites will have neither, either because they are a territory,
an international delegation or a place that only has offices (eg a DFID or UKTI
office). However, there will always be contact details for one or more HQs in a
UK in the world site.'