Remove dependent associations on object destruction.
We're going to be removing some "dummy" data from production shortly. We
want to make sure we can do this through our models so we're starting to
go through and add the relevant "dependent: :destroy" options.

With this commit, destroying a document will result in the associated
document_attachments and document_relations being removed. We've added a
missing test to ensure that destroying a document also removes the
document_authors relationship.