Add script to remove "dummy" documents in production.
We've been making commits during the day that should allow us to destroy
a DocumentIdentity and have it destroy all its dependencies.

We've tested the script in development using the production data and are
confident that it's Doing The Right Thing.

The delete_documents_for_deletion.sql script contains the queries we
were using in development to test that the script was removing what we
expected it to remove. To do this, we first took a mysqldump of the
database, ran the script, took another dump and then grep'd the diff of
the dumps to detect the number of records removed and compared that
number to the output from the sql statements in
delete_documents_for_deletion.sql. *NOTE* The '--skip-extended-insert'
flag gives us one line per record which facilitates diffing the output.

 $ mysqldump whitehall_development --skip-extended-insert > before.sql
 $ script/rails runner script/delete_documents_for_deletion.rb
 $ mysqldump whitehall_development --skip-extended-insert > after.sql
 $ diff {before,after}.sql > whitehall.sql.diff
 $ grep "INSERT INTO \`document_authors\`" whitehall.sql.diff | wc -l
 $ # More grepping...