Set the correct content-type of our pdf thumbnails.
CarrierWave assumes that the content type of any versions will not be
different from that of the original uploaded file. When uploading to S3,
it sets this as the content type of the file being stored in the bucket.
S3 then serves the file with that content type.

Some browsers are not happy when trying to display a PNG thumbnail which
has been served with the content type "application/pdf", including
Safari. This commit should allow us to fix existing and future
thumbnails so they are served by S3 with the correct content type, and
so display properly in Safari.