Remove .rbenv-version.
The presence of this file is a pain for people switching between
1.9.2-p290 and 1.9.3-p0-perf. Developing against 1.9.3-p0 is much faster
than 1.9.2. We've updated the README to explain that we've tested the
app successfully against these two versions.

This file was originally a part of the repository to mirror the version
used on production servers, but at the moment the benefits of
consistency of ruby versions are outweighed by those of a much faster
development environment.