Fix the search indexing of documents.
We identified two problems with the previous implementation: publishing
a new edition of a document and deleting a new draft of a published
document would result in the documents being removed from the search
index. This was because of the `unindex_after: [:archive, :delete]`
option passed to `searchable`. Our solution is to manually trigger an
addition or removal from the index as required, after a document is
published, archived or deleted.