Ensure the First Secretary of State is ordered properly in production.
"First Secretary of State" isn't really a role, so we can't use its name directly to order.
Ideally we'd model which minister was also the "First Secretary", but that introduces more
problems. If we'd made "First Secretary of State" a ministerial role, William Hague would appear
twice in the list of cabinet members. If we'd made it a non-ministerial role, it wouldn't have
been sortable in the set that MinisterialRole.cabinet returned.

This is a quick, pragmatic solution, but more thought probably needs to go into whether or not
the lists of ministers are lists of roles, or should instead be lists of people (along with the
roles they currently have).