Allow editors to feature news articles for a specific country.
I added some tests around document_organisations_controller and
organsations_controller (testing the featuring/unfeaturing and for the
presence of the featuring/unfeaturing buttons). I then copied these to
test the new document_countries_controller and new functionality in the
countries_controller.

This commit only adds the admin functionality, @lazyatom is working on
the frontend visual of the featured news for a country. We should be
able to copy/amend the administering-organisations.feature once we have
the frontend visuals in place.