Add API for consultations.
This uses RABL (see https://github.com/nesquena/rabl) and provides the basis
for adding API endpoints to other types of document, too: both the consultation
and the consultation response already extend a shared Document RABL template.

This is only a provisional API - as we build it into other parts of Whitehall,
we'll hopefully get a clearer idea of what it should look like - this should
also be co-ordinated with the citizen APIs.

For now, we're still beta, so we can change.