Remove redundant migration
The presence of this migration was preventing db:migrate running from a
clean database.

The migration was added in order to remove the ranks table from
preview/production. The ranks table was originally added in commit
13c75e. That commit was reverted in 36e5ec but the migration had already
been run in preview/production. I then added this "temporary" migration
in 59af29 but neglected to remove it after running it on
preview/production.