Remove jquery-ui.
The static app provides/includes its own version of jquery-ui so we can
remove ours.  Although the filenames are the same, diffing the two
versions did show some differences.  I'm going to assume that we'll spot
any breakages in the app should any of those differences turn out to be
important.