An interrupt exception (Ctrl-C) should stop the tests running.
The monkey-patch in ActiveSupport for MiniTest is out-of-date (at least
for the version of Ruby I am running). This means that currently when
you try to interrupt a test, the current test is marked as having an
error, but execution continues.

Given this patch is a monkey-patch on top of a monkey-patch, it's clearly
not a great solution, but it solves my immediate problem for now and
it would be a fair bit of work to sort out what should happen for all
the combinations of MiniTest & ActiveSupport versions.