Rename document to edition in functional tests.
I've changed the variables that were obvious candidates for being
renamed from document to edition but have left some tests talking about
documents.  This is because the concept of editions should be pretty
much invisible to normal users of the system; they should only be
concerned with documents.  This situation is slightly confusing because
of the overlap between a document and its published edition.