Rename `:document` parameter key in admin controllers to `:edition`.
These controllers are really dealing with editions and not documents, so
it makes things clearer if we use a more appropriately named key.

This has some consequences for IDs and names of HTML form-related
elements on the admin pages, because these are generated by Rails helper
methods. I believe I have found all the references to these IDs and
names in the CSS, but there's a chance I've missed something. Apologies
if I have.