Remove custom admin show templates for speech & consultation response.
Having a single show template has a number of benefits which I believe
outweigh the additional complexity of conditional within that template.

The principle benefit is that all edition types will benefit from any common
workflow/attribute changes (such as the audit history) automatically, unless
they specifically opt out. Previous to this commit, version history was not
visible for speeches, for example.

A secondary benefit is that improvements to the layout of the admin preview
will also be shared across all edition types. This is something I'm
considering independently.

Finally, I think it might be true that removing conditionals in templates in
this way results in more tests; if there's only a single generic show
template, then we might be able to test it once, rather than exercising it
through all the various `Admin::EditionController` subclasses using shared
tests.