Move admin fact check request routes to be under edition.
Because the routes are shallow, the only URL path that changes is the
one for the create action. And since this is only used internally, we
don't need to worry about redirecting the old URL. Note that
specifically I was concerned about changes to the link used in the fact
check request notification email, but that only uses the URL for the
edit action which does not change in this commit.