Improve the layout of the admin edition preview page.
All other admin pages still use the old interface, although I will begin to
migrate them soon.

When I added creation of editorial remarks to this page, the layout got worse;
this has been an increasing trend as there has been no focus on the UI of the
admin interface.

I've decided to adopt an external framework for the admin interface, which
should allow us to add new features to it in a more sustainable way.