Comment out broken cucumber step!!
Whilst working with Mark to fix his build, we realised that this step doesn't test anything.  .document doesn't match anything, and if it is changed to .policy, it fails because PolicyTopic doesn't have an editions method!  I have commented this out to let Mark continue.  Will work out the intention and fix in a later commit.