Do not attempt to send an email to a blank email address.
This was a particular issue, because if you edit your account settings
and don't enter an email address, it gets saved as an empty String and
not as nil. There was a check in the notification code, but it was only
checking for nil email addresses and not empty String.