Organize media-player javascript more sensibly.
Move the media-player javascript into the common directory. Both
admin.js and application.js include everything in this directory by
default. This means there is no longer any need for explicit javascript
include tags for the media-player javascript and I believe no need for
it to be listed in the assets which need pre-compiling.